Refactoring some classes out of grpc_observability/__init__.py
